Plenty of changes have been applied. Show the details below.
The vacation request process have got the conception of consistent vacation acceptance. Now it has 4 stages, and manager and admin obligations have been splitted. Admins accept the request on the 1st and 3rd stages, managers on the 2nd stage.  Nowadays only the first stage is finally completed, others are in development.
Also we have other improvements. Firstly, the repositories have got more flex via the adding of search delegate to getters. Finally, the page with list of users for admins have been restored and filter functions on it have been partly recovered.